Interdiction de supprimer un fichier possédant un id permanent.

+++ module/Application/src/Application/Service/Fichier/FichierService.php
@@ -183,6 +183,12 @@ class FichierService extends BaseService
         $this->entityManager->beginTransaction();
         try {
             foreach ($fichiers as $fichier) {
+                if ($fichier->getIdPermanent() !== null) {
+                    throw new Exception(sprintf(
+                        "Interdit de supprimer un fichier possédant un id permanent (en l'occurence : '%s')",
+                        $fichier->getIdPermanent()
+                    ));
+                }
                 $filePaths[] = $this->computeDestinationFilePathForFichier($fichier);
                 $this->entityManager->remove($fichier);
                 $this->entityManager->flush($fichier);
